Refactor event model package structure for better organization
Moved event-related classes to a unified "events" package structure under domain.model. This better follows naming standards for packages -- avoid camel case -- and makes it clear that these classes belong together.

# Conflicts:
#	src/test/kotlin/no/nb/mlt/wls/domain/WLSServiceTest.kt
