+++ src/main/kotlin/no/nb/mlt/wls/domain/model/events/Event.kt
@@ -1,4 +1,4 @@
-package no.nb.mlt.wls.domain.model
+package no.nb.mlt.wls.domain.model.events
 
 interface Event {
     val id: String
