+++ src/main/kotlin/no/nb/mlt/wls/domain/model/events/catalog/CatalogEvent.kt
@@ -1,6 +1,6 @@
-package no.nb.mlt.wls.domain.model.catalogEvents
+package no.nb.mlt.wls.domain.model.events.catalog
 
-import no.nb.mlt.wls.domain.model.Event
+import no.nb.mlt.wls.domain.model.events.Event
 import java.time.Instant
 
 sealed interface CatalogEvent : Event {
