+++ src/main/kotlin/no/nb/mlt/wls/domain/model/events/storage/OrderDeleted.kt
@@ -1,4 +1,4 @@
-package no.nb.mlt.wls.domain.model.storageEvents
+package no.nb.mlt.wls.domain.model.events.storage
 
 import no.nb.mlt.wls.domain.model.HostName
 import java.util.UUID
