+++ src/main/kotlin/no/nb/mlt/wls/domain/model/events/storage/StorageEvent.kt
@@ -1,0 +1,5 @@
+package no.nb.mlt.wls.domain.model.events.storage
+
+import no.nb.mlt.wls.domain.model.events.Event
+
+sealed interface StorageEvent : Event
