--- src/main/kotlin/no/nb/mlt/wls/domain/model/storageEvents/StorageEvent.kt
@@ -1,5 +1,0 @@
-package no.nb.mlt.wls.domain.model.storageEvents
-
-import no.nb.mlt.wls.domain.model.Event
-
-sealed interface StorageEvent : Event
