+++ src/main/kotlin/no/nb/mlt/wls/domain/ports/outbound/EventProcessor.kt
@@ -1,6 +1,6 @@
 package no.nb.mlt.wls.domain.ports.outbound
 
-import no.nb.mlt.wls.domain.model.Event
+import no.nb.mlt.wls.domain.model.events.Event
 
 fun interface EventProcessor<in T : Event> {
     suspend fun handleEvent(event: T)
