+++ src/main/kotlin/no/nb/mlt/wls/infrastructure/repositories/event/MongoCatalogEventRepositoryAdapter.kt
@@ -2,7 +2,7 @@ package no.nb.mlt.wls.infrastructure.repositories.event
 
 import io.github.oshai.kotlinlogging.KotlinLogging
 import kotlinx.coroutines.reactor.awaitSingle
-import no.nb.mlt.wls.domain.model.catalogEvents.CatalogEvent
+import no.nb.mlt.wls.domain.model.events.catalog.CatalogEvent
 import no.nb.mlt.wls.domain.ports.outbound.EventRepository
 import no.nb.mlt.wls.domain.ports.outbound.RepositoryException
 import org.springframework.data.mongodb.repository.Query
